fix: move react-dom/server to Puppeteer generator file
- Move HTML generation directly into puppeteer-generator.ts
- Import react-dom/server in the generator file (allowed in server-side code)
- Remove separate API route that was causing issues
- This should fix the Netlify build error while maintaining functionality
